Add pseudo n2 linear fetch to 3D fortran wrapper

+++ wrappers/Fortran/mui_f_wrapper_1d.f90
@@ -12498,21 +12498,21 @@ module mui_1d_f
     end subroutine mui_set_forget_length_1fx_f
 
     subroutine mui_set_forget_length_1d_f(uniface,length) bind(C)
-      import :: c_ptr,c_float
+      import :: c_ptr,c_double
       type(c_ptr), intent(in), value :: uniface
-      real(kind=c_float), intent(in), value :: length
+      real(kind=c_double), intent(in), value :: length
     end subroutine mui_set_forget_length_1d_f
 
     subroutine mui_set_forget_length_1dx_f(uniface,length) bind(C)
-      import :: c_ptr,c_float
+      import :: c_ptr,c_double
       type(c_ptr), intent(in), value :: uniface
-      real(kind=c_float), intent(in), value :: length
+      real(kind=c_double), intent(in), value :: length
     end subroutine mui_set_forget_length_1dx_f
 
     subroutine mui_set_forget_length_1t_f(uniface,length) bind(C)
-      import :: c_ptr,c_float
+      import :: c_ptr,c_double
       type(c_ptr), intent(in), value :: uniface
-      real(kind=c_float), intent(in), value :: length
+      real(kind=c_double), intent(in), value :: length
     end subroutine mui_set_forget_length_1t_f
 
     !******************************************
